Fix size types in BoringSSL code
BN_num_bytes() and RSA_size() return "unsigned int" in BoringSSL
while in OpenSSL they use regular (signed) "int". Update the code
to expect correct types from these functions.

+++ src/soter/boringssl/soter_ec_key.c
@@ -95,8 +95,8 @@ static char* ec_priv_key_tag(int curve)
 
 static size_t bn_encode(const BIGNUM* bn, uint8_t* buffer, size_t length)
 {
-    int bn_size = BN_num_bytes(bn);
-    if (length < (size_t)bn_size) {
+    unsigned bn_size = BN_num_bytes(bn);
+    if (length < bn_size) {
         return 0;
     }
     memset(buffer, 0, length - bn_size);
